+++ apps/mcp-server/RELEASES.md
@@ -7,3 +7,11 @@ Based on:
 - Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
 ### Generated
 - [mcp-typescript v0.0.1] apps/mcp-server
+
+## 2026-05-29 01:16:45
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.768.2 (2.889.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [mcp-typescript v0.0.2] apps/mcp-server